+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Replaced all unsafe code from svenz-rust2 crate
+- Replaced all unsafe code from sevenz-rust2 and lzma-rust2 with safe alternatives
 - sha2 is now an optional dependency that is activated when using the aes256 features
 
 ### 0.7.0 - 2025-02-24